save intermediate results

+++ src/main/java/fund/data/assets/dto/asset/exchange/FixedRateBondPartialSellDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 import java.util.Map;
 
 /**
- * DTO для обслуживания внесения в систему данных о продаже части облигаций в пакет бумаг данного выпуска облигаций.
+ * DTO для обслуживания внесения в систему данных о продаже части облигаций из пакета бумаг выпуска облигаций.
  * @version 0.0.1-alpha
  * @author MarkDementev a.k.a JavaMarkDem
  */
@@ -18,5 +18,5 @@ import java.util.Map;
 @AllArgsConstructor
 public class FixedRateBondPartialSellDTO {
     @NotNull
-    private Map<String, Float> assetOwnersWithAssetCountsToSell;
+    private Map<String, Integer> assetOwnersWithAssetCountsToSell;
 }
